Add buffer types for the count swap shader
- Encapsulate workgroup grid dimensions calculations into a new SwapPass enum

+++ image_annealing_shaders/src/constant/count_swap.rs
@@ -1,7 +1,7 @@
 use crate::compute::WorkgroupDimensions;
 use std::io::Write;
 
-pub const N_CHANNEL: u32 = 4;
+pub const N_CHANNEL: usize = 4;
 
 pub fn n_channel<W: Write>(mut writer: W) -> std::io::Result<()> {
     writeln!(writer, "let n_channel: u32 = {}u;", N_CHANNEL)
